Replace int with long long int

+++ src/number_theory/divisor.cpp
@@ -8,9 +8,9 @@
 
 #include <vector>
 
-std::vector<int> divisor(int n) {
-  std::vector<int> res;
-  for (int i = 1; i * i <= n; i++) {
+std::vector<long long int> divisor(long long int n) {
+  std::vector<long long int> res;
+  for (long long int i = 1; i * i <= n; i++) {
     if (n % i == 0) {
       res.push_back(i);
       if (i != n / i)
