+++ src/number_theory/is_prime.cpp
@@ -7,8 +7,8 @@
         n == 1 の場合は例外
 */
 
-bool is_prime(int n) {
-  for (int i = 2; i * i <= n; i++) {
+bool is_prime(long long int n) {
+  for (long long int i = 2; i * i <= n; i++) {
     if (n % i == 0)
       return false;
   }
